+++ CHANGELOG.md
@@ -4,8 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Fix: Add `MastercardIcon` on `PaymentButton`.
-- Fix: Add new `MastercardIcon` component.
+- Fix: Add `KarlPaymentButtonVirement`, `KarlPaymentButtonMastercard`,
+  `KarlPaymentButtonCb` and `KarlPaymentButtonVirement` on `PaymentButton`.
+- Fix: Add new `MastercardIcon` and `CbIcon` component.
 - Fix: Propagate `className` prop to `Grid` and `GridCol` components.
 - Fix: Remove `error` label on Checkbox and RadioButton guideline example.
 - Feature: Add `Separator` component.